Fixed problems found with gunowners.org caused by more than 400 entries on the page. Upgraded system to work with up to 1000 endorsements on a page. Preventing duplicate candidates from being added to a Voter Guide Possibility. Added "started by" to the Voter Guide Possibility page if there is a value. Fixed problem with duplicate positions not being deleted.

+++ voter_guide/models.py
@@ -115,71 +115,22 @@ WEBSITES_WE_DO_NOT_SCAN_FOR_ENDORSEMENTS = [
     # 'wapo.st', 'westword.com', 'wsj.com',
     # 'youtu.be', 'youtube.com',
 ]
-POSSIBLE_ENDORSEMENT_NUMBER_LIST = [
-    "001", "002", "003", "004", "005", "006", "007", "008", "009", "010",
-    "011", "012", "013", "014", "015", "016", "017", "018", "019", "020",
-    "021", "022", "023", "024", "025", "026", "027", "028", "029", "030",
-    "031", "032", "033", "034", "035", "036", "037", "038", "039", "040",
-    "041", "042", "043", "044", "045", "046", "047", "048", "049", "050",
-    "051", "052", "053", "054", "055", "056", "057", "058", "059", "060",
-    "061", "062", "063", "064", "065", "066", "067", "068", "069", "070",
-    "071", "072", "073", "074", "075", "076", "077", "078", "079", "080",
-    "081", "082", "083", "084", "085", "086", "087", "088", "089", "090",
-    "091", "092", "093", "094", "095", "096", "097", "098", "099", "100",
-    "101", "102", "103", "104", "105", "106", "107", "108", "109", "110",
-    "111", "112", "113", "114", "115", "116", "117", "118", "119", "120",
-    "121", "122", "123", "124", "125", "126", "127", "128", "129", "130",
-    "131", "132", "133", "134", "135", "136", "137", "138", "139", "140",
-    "141", "142", "143", "144", "145", "146", "147", "148", "149", "150",
-    "151", "152", "153", "154", "155", "156", "157", "158", "159", "160",
-    "161", "162", "163", "164", "165", "166", "167", "168", "169", "170",
-    "171", "172", "173", "174", "175", "176", "177", "178", "179", "180",
-    "181", "182", "183", "184", "185", "186", "187", "188", "189", "190",
-    "191", "192", "193", "194", "195", "196", "197", "198", "199", "200"]
+POSSIBILITY_LIST_LIMIT = 1000  # Limit to 1000 to avoid very slow page loading, formerly 200, then 400
+# We want a list of 1000 numbers converted to strings, so we use the last 4 digits
+max_number_of_digits = len(str(POSSIBILITY_LIST_LIMIT))
+leading_zeros = ""
+for i in range(max_number_of_digits):
+    leading_zeros += "0"
+number_list_temp = []
+for i in range(POSSIBILITY_LIST_LIMIT):
+    digit = i + 1
+    string_temp = leading_zeros + str(digit)
+    number_list_temp.append(string_temp[-max_number_of_digits:])
+POSSIBLE_ENDORSEMENT_NUMBER_LIST = number_list_temp
 
 # Since there are some scenarios where we remove candidates where positions are already stored, we need to review more
 #  positions than we are prepared to store
-POSSIBLE_ENDORSEMENT_NUMBER_LIST_FULL = [
-    "001", "002", "003", "004", "005", "006", "007", "008", "009", "010",
-    "011", "012", "013", "014", "015", "016", "017", "018", "019", "020",
-    "021", "022", "023", "024", "025", "026", "027", "028", "029", "030",
-    "031", "032", "033", "034", "035", "036", "037", "038", "039", "040",
-    "041", "042", "043", "044", "045", "046", "047", "048", "049", "050",
-    "051", "052", "053", "054", "055", "056", "057", "058", "059", "060",
-    "061", "062", "063", "064", "065", "066", "067", "068", "069", "070",
-    "071", "072", "073", "074", "075", "076", "077", "078", "079", "080",
-    "081", "082", "083", "084", "085", "086", "087", "088", "089", "090",
-    "091", "092", "093", "094", "095", "096", "097", "098", "099", "100",
-    "101", "102", "103", "104", "105", "106", "107", "108", "109", "110",
-    "111", "112", "113", "114", "115", "116", "117", "118", "119", "120",
-    "121", "122", "123", "124", "125", "126", "127", "128", "129", "130",
-    "131", "132", "133", "134", "135", "136", "137", "138", "139", "140",
-    "141", "142", "143", "144", "145", "146", "147", "148", "149", "150",
-    "151", "152", "153", "154", "155", "156", "157", "158", "159", "160",
-    "161", "162", "163", "164", "165", "166", "167", "168", "169", "170",
-    "171", "172", "173", "174", "175", "176", "177", "178", "179", "180",
-    "181", "182", "183", "184", "185", "186", "187", "188", "189", "190",
-    "191", "192", "193", "194", "195", "196", "197", "198", "199", "200",
-    "201", "202", "203", "204", "205", "206", "207", "208", "209", "210",
-    "211", "212", "213", "214", "215", "216", "217", "218", "219", "220",
-    "221", "222", "223", "224", "225", "226", "227", "228", "229", "230",
-    "231", "232", "233", "234", "235", "236", "237", "238", "239", "240",
-    "241", "242", "243", "244", "245", "246", "247", "248", "249", "250",
-    "251", "252", "253", "254", "255", "256", "257", "258", "259", "260",
-    "261", "262", "263", "264", "265", "266", "267", "268", "269", "270",
-    "271", "272", "273", "274", "275", "276", "277", "278", "279", "280",
-    "281", "282", "283", "284", "285", "286", "287", "288", "289", "290",
-    "291", "292", "293", "294", "295", "296", "297", "298", "299", "300",
-    "301", "302", "303", "304", "305", "306", "307", "308", "309", "310",
-    "311", "312", "313", "314", "315", "316", "317", "318", "319", "320",
-    "321", "322", "323", "324", "325", "326", "327", "328", "329", "330",
-    "331", "332", "333", "334", "335", "336", "337", "338", "339", "340",
-    "341", "342", "343", "344", "345", "346", "347", "348", "349", "350",
-    "351", "352", "353", "354", "355", "356", "357", "358", "359", "360",
-    "361", "362", "363", "364", "365", "366", "367", "368", "369", "370",
-    "371", "372", "373", "374", "375", "376", "377", "378", "379", "380",
-    "381", "382", "383", "384", "385", "386", "387", "388", "389", "390",
-    "391", "392", "393", "394", "395", "396", "397", "398", "399", "400"]
+# POSSIBLE_ENDORSEMENT_NUMBER_LIST_FULL = []
 
 
 class VoterGuideManager(models.Manager):
@@ -2929,22 +2880,39 @@ class VoterGuidePossibilityManager(models.Manager):
         return False
 
     def delete_voter_guide_possibility(self, voter_guide_possibility_id):
+        status = ""
+        success = True
+
         voter_guide_possibility_id = convert_to_int(voter_guide_possibility_id)
         voter_guide_possibility_deleted = False
 
         try:
             if positive_value_exists(voter_guide_possibility_id):
+                queryset = VoterGuidePossibilityPosition.objects.all()
+                queryset = queryset.filter(voter_guide_possibility_parent_id=voter_guide_possibility_id)
+                number_deleted = queryset.delete()
+                status += str(number_deleted) + "VoterGuidePossibilityPosition_DELETED "
+
+        except Exception as e:
+            handle_exception(e, logger=logger)
+            success = False
+
+        try:
+            if success and positive_value_exists(voter_guide_possibility_id):
                 results = self.retrieve_voter_guide_possibility(voter_guide_possibility_id)
                 if results['voter_guide_possibility_found']:
                     voter_guide_possibility = results['voter_guide_possibility']
                     voter_guide_possibility_id = voter_guide_possibility.id
                     voter_guide_possibility.delete()
+                    status += "VOTER_GUIDE_POSSIBILITY_DELETED "
                     voter_guide_possibility_deleted = True
         except Exception as e:
             handle_exception(e, logger=logger)
+            success = False
 
         results = {
-            'success':                          voter_guide_possibility_deleted,
+            'success':                          success,
+            'status':                           status,
             'voter_guide_possibility_deleted':  voter_guide_possibility_deleted,
             'voter_guide_possibility_id':       voter_guide_possibility_id,
         }
